fix: isolate types for cleanliness and lightness
